의존성 주입 테스트
1. 의존성 주입 테스트
싱글톤을 사용할 경우 API에서 동시다발적으로 Service의 인스턴스를 요청할 경우 예상치 못한 에러가 발생할 수 있음.
